Extract info helpers into example instead of src
