+++ examples/extensions/helpers/config.extractors.ts
